refactor(protocol): extract group_mesh tests into dedicated module
Move 82 tests (3,400+ lines) from group_mesh.rs into group_mesh_tests.rs,
reducing the main file from 4,818 to 1,408 lines. Adds shared test helpers
and bumps a few internal items to pub(crate) for cross-module test access.
